+++ CHANGES.rst
@@ -197,6 +197,8 @@ Unreleased
 - Remove external links indicator
 - Improve visual appearance of admonition components
 - Add new ``cloud-docs`` documentation project
+- Switch to modern responsive 3-column layout theme ``sphinx-basic-ng``.
+  Thanks, @pradyunsg.
 
 
 2023/05/15 0.27.1